+++ changelog.md
@@ -5,6 +5,8 @@
 > [!IMPORTANT]
 > This version is not released yet and is under active development.
 
+- Fix `gha-utils update-exclude-newer` to handle relative date strings.
+
 ## [5.5.1 (2026-01-30)](https://github.com/kdeldycke/workflows/compare/v5.5.0...v5.5.1)
 
 - Replace workflow shell scripts with CLI commands.